+++ espresso/src/org.graalvm.continuations/src/org/graalvm/continuations/Generator.java
@@ -64,8 +64,8 @@ public abstract class Generator<E> implements Enumeration<E>, Serializable {
      */
     @SuppressWarnings("this-escape")
     protected Generator() {
-        continuation = new Continuation((EntryPoint & Serializable) suspendCapability -> {
-            this.suspendCapability = suspendCapability;
+        continuation = new Continuation((EntryPoint & Serializable) cap -> {
+            this.suspendCapability = cap;
             generate();
         });
     }